MY_REPOS/Lambda-Resource-Static-Assets/2-content/14-Pure-Education/java-data-struc/java-prep/src/test/java/data/structures/java/strings/LongestStringWithNUniqueTest.java
package data.structures.java.strings;

import org.junit.Test;

import static org.junit.Assert.*;

public class LongestStringWithNUniqueTest
{

  @Test
  public void longestSubstringLengthWith2Unique()
  {
    //"abcbbbbcccbdddadacb" -> "bcbbbbcccb"
    assertEquals(10, LongestStringWithNUnique.longestSubstringLengthWithNUnique("abcbbbbcccbdddadacb", 2));
  }

  @Test
  public void longestSubstringWithNUnique()
  {
    assertEquals("bcbbbbcccb", LongestStringWithNUnique.longestSubstringWithNUnique("abcbbbbcccbdddadacb", 2));
    assertEquals("xxxyyy", LongestStringWithNUnique.longestSubstringWithNUnique("abcbbxxxyyybdddadacb", 2));
    assertEquals("cadcacacaca", LongestStringWithNUnique.longestSubstringWithNUnique("abcadcacacaca", 3));
  }


}